convert scalars from float32 to float64

+++ canvas.go
@@ -57,7 +57,7 @@ func (c *Canvas) NewPPM() *PPM {
 		for x := 0; x < c.Width; x++ {
 			color := c.Grid[x][y]
 			// Iteratve over RGB values of this pixel
-			for _, valueFloat32 := range []float32{color.Red, color.Green, color.Blue} {
+			for _, valueFloat32 := range []float64{color.Red, color.Green, color.Blue} {
 
 				// Scale the color value
 				valueInt := int(math.Round((float64)(valueFloat32 * 255)))
